+++ CHANGELOG.md
@@ -12,6 +12,7 @@ New features:
 - As we have a new language, added an AIFPL syntax highlighter (this makes the docs render correctly).
 - Removed Claude Sonnet 3.5 support as that is now deprecated by Anthropic.
 - Removed the calculator tool.  The LLMs now have AIFPL so there's no need to carry two tools that can do calculations.
+- Added Claude Sonnet 4.5 support.
 
 Bug fixes: